sua hoc truc tuyen

+++ Backend/Backend/Models/ClassEnrollments.cs
@@ -8,12 +8,13 @@
 
         public int StudentId { get; set; }
 
-        public int Status { get; set; }
+        public byte Status { get; set; }
 
         public DateTime EnrolledDate { get; set; }
 
         public TeacherClass? TeacherClass { get; set; }
 
         public User? Student { get; set; }
+        
     }
 }
